Move all non-dimension stuff out of HotMDimensions
